refactor: yggctl actions into non-anonymous functions
The purpose of this refactor is to follow better design patterns
and move the yggctl anonymous action functions into
non-anonymous functions defined in a separate file to
prevent polluting the yggctl main file.
