+++ cmd/yggctl/main.go
@@ -1,19 +1,12 @@
 package main
 
 import (
-	"encoding/json"
 	"fmt"
-	"io"
 	"os"
-	"text/tabwriter"
 
 	"git.sr.ht/~spc/go-log"
 
-	"github.com/godbus/dbus/v5"
-	"github.com/google/uuid"
-	"github.com/redhatinsights/yggdrasil"
 	"github.com/redhatinsights/yggdrasil/internal/constants"
-	"github.com/redhatinsights/yggdrasil/ipc"
 	"github.com/urfave/cli/v2"
 )
 
@@ -69,21 +62,7 @@ func main() {
 							Usage:    "set directive to `STRING`",
 						},
 					},
-					Action: func(c *cli.Context) error {
-						var metadata map[string]string
-						if err := json.Unmarshal([]byte(c.String("metadata")), &metadata); err != nil {
-							return cli.Exit(fmt.Errorf("cannot unmarshal metadata: %w", err), 1)
-						}
-
-						data, err := generateMessage("data", c.String("response-to"), c.String("directive"), c.Args().First(), metadata, c.Int("version"))
-						if err != nil {
-							return cli.Exit(fmt.Errorf("cannot marshal message: %w", err), 1)
-						}
-
-						fmt.Println(string(data))
-
-						return nil
-					},
+					Action: generateDataMessageAction,
 				},
 				{
 					Name:    "control-message",
@@ -108,16 +87,7 @@ func main() {
 							Usage:    "set message type to `STRING`",
 						},
 					},
-					Action: func(c *cli.Context) error {
-						data, err := generateMessage(c.String("type"), c.String("response-to"), "", c.Args().First(), nil, c.Int("version"))
-						if err != nil {
-							return cli.Exit(fmt.Errorf("cannot marshal message: %w", err), 1)
-						}
-
-						fmt.Println(string(data))
-
-						return nil
-					},
+					Action: generateControlMessageAction,
 				},
 			},
 		},
@@ -136,53 +106,7 @@ func main() {
 							Value: "text",
 						},
 					},
-					Action: func(c *cli.Context) error {
-						var conn *dbus.Conn
-						var err error
-
-						if os.Getenv("DBUS_SESSION_BUS_ADDRESS") != "" {
-							conn, err = dbus.ConnectSessionBus()
-						} else {
-							conn, err = dbus.ConnectSystemBus()
-						}
-						if err != nil {
-							return cli.Exit(fmt.Errorf("cannot connect to bus: %w", err), 1)
-						}
-
-						obj := conn.Object("com.redhat.Yggdrasil1", "/com/redhat/Yggdrasil1")
-						var workers map[string]map[string]string
-						if err := obj.Call("com.redhat.Yggdrasil1.ListWorkers", dbus.Flags(0)).Store(&workers); err != nil {
-							return cli.Exit(fmt.Errorf("cannot list workers: %v", err), 1)
-						}
-
-						switch c.String("format") {
-						case "json":
-							data, err := json.Marshal(workers)
-							if err != nil {
-								return cli.Exit(fmt.Errorf("cannot marshal workers: %v", err), 1)
-							}
-							fmt.Println(string(data))
-						case "table":
-							writer := tabwriter.NewWriter(os.Stdout, 4, 4, 2, ' ', 0)
-							fmt.Fprintf(writer, "WORKER\tFIELD\tVALUE\n")
-							for worker, features := range workers {
-								for field, value := range features {
-									fmt.Fprintf(writer, "%v\t%v\t%v\n", worker, field, value)
-								}
-								_ = writer.Flush()
-							}
-						case "text":
-							for worker, features := range workers {
-								for field, value := range features {
-									fmt.Printf("%v - %v: %v\n", worker, field, value)
-								}
-							}
-						default:
-							return cli.Exit(fmt.Errorf("unknown format type: %v", c.String("format")), 1)
-						}
-
-						return nil
-					},
+					Action: workersAction,
 				},
 			},
 		},
@@ -205,50 +129,7 @@ func main() {
 					Value:   "{}",
 				},
 			},
-			Action: func(c *cli.Context) error {
-				var conn *dbus.Conn
-				var err error
-
-				if os.Getenv("DBUS_SESSION_BUS_ADDRESS") != "" {
-					conn, err = dbus.ConnectSessionBus()
-				} else {
-					conn, err = dbus.ConnectSystemBus()
-				}
-				if err != nil {
-					return cli.Exit(fmt.Errorf("cannot connect to bus: %w", err), 1)
-				}
-
-				var metadata map[string]string
-				if err := json.Unmarshal([]byte(c.String("metadata")), &metadata); err != nil {
-					return cli.Exit(fmt.Errorf("cannot unmarshal metadata: %w", err), 1)
-				}
-
-				var data []byte
-				var r io.Reader
-				if c.Args().First() == "-" {
-					r = os.Stdin
-				} else {
-					r, err = os.Open(c.Args().First())
-				}
-				if err != nil {
-					return cli.Exit(fmt.Errorf("cannot open file for reading: %w", err), 1)
-				}
-				data, err = io.ReadAll(r)
-				if err != nil {
-					return cli.Exit(fmt.Errorf("cannot read data: %w", err), 1)
-				}
-
-				id := uuid.New().String()
-
-				obj := conn.Object("com.redhat.Yggdrasil1", "/com/redhat/Yggdrasil1")
-				if err := obj.Call("com.redhat.Yggdrasil1.Dispatch", dbus.Flags(0), c.String("worker"), id, metadata, data).Store(); err != nil {
-					return cli.Exit(fmt.Errorf("cannot dispatch message: %w", err), 1)
-				}
-
-				fmt.Printf("Dispatched message %v to worker %v\n", id, c.String("worker"))
-
-				return nil
-			},
+			Action: dispatchAction,
 		},
 		{
 			Name:        "listen",
@@ -262,71 +143,11 @@ func main() {
 					Required: true,
 				},
 			},
-			Action: func(ctx *cli.Context) error {
-				var conn *dbus.Conn
-				var err error
-
-				if os.Getenv("DBUS_SESSION_BUS_ADDRESS") != "" {
-					conn, err = dbus.ConnectSessionBus()
-				} else {
-					conn, err = dbus.ConnectSystemBus()
-				}
-				if err != nil {
-					return cli.Exit(fmt.Errorf("cannot connect to bus: %w", err), 1)
-				}
-
-				if err := conn.AddMatchSignal(); err != nil {
-					return cli.Exit(fmt.Errorf("cannot add match signal: %w", err), 1)
-				}
-
-				signals := make(chan *dbus.Signal)
-				conn.Signal(signals)
-				for s := range signals {
-					switch s.Name {
-					case "com.redhat.Yggdrasil1.WorkerEvent":
-						worker, ok := s.Body[0].(string)
-						if !ok {
-							return cli.Exit(fmt.Errorf("cannot cast %T as string", s.Body[0]), 1)
-						}
-						name, ok := s.Body[1].(uint32)
-						if !ok {
-							return cli.Exit(fmt.Errorf("cannot cat %T as uint32", s.Body[1]), 1)
-						}
-						var message string
-						if len(s.Body) > 2 {
-							message, ok = s.Body[2].(string)
-							if !ok {
-								return cli.Exit(fmt.Errorf("cannot cast %T as string", s.Body[0]), 1)
-							}
-						}
-						log.Printf("%v: %v: %v", worker, ipc.WorkerEventName(name), message)
-
-					}
-				}
-				return nil
-			},
+			Action: listenAction,
 		},
 	}
 
-	app.Action = func(c *cli.Context) error {
-		if c.Bool("generate-man-page") || c.Bool("generate-markdown") {
-			type GenerationFunc func() (string, error)
-			var generationFunc GenerationFunc
-			if c.Bool("generate-man-page") {
-				generationFunc = c.App.ToMan
-			} else if c.Bool("generate-markdown") {
-				generationFunc = c.App.ToMarkdown
-			}
-			data, err := generationFunc()
-			if err != nil {
-				return err
-			}
-			fmt.Println(data)
-			return nil
-		}
-
-		return cli.ShowAppHelp(c)
-	}
+	app.Action = generateManPage
 	app.EnableBashCompletion = true
 
 	if err := app.Run(os.Args); err != nil {
@@ -334,29 +155,22 @@ func main() {
 	}
 }
 
-func generateMessage(messageType, responseTo, directive, content string, metadata map[string]string, version int) ([]byte, error) {
-	switch messageType {
-	case "data":
-		msg, err := generateDataMessage(yggdrasil.MessageType(messageType), responseTo, directive, []byte(content), metadata, version)
-		if err != nil {
-			return nil, err
+func generateManPage(c *cli.Context) error {
+	if c.Bool("generate-man-page") || c.Bool("generate-markdown") {
+		type GenerationFunc func() (string, error)
+		var generationFunc GenerationFunc
+		if c.Bool("generate-man-page") {
+			generationFunc = c.App.ToMan
+		} else if c.Bool("generate-markdown") {
+			generationFunc = c.App.ToMarkdown
 		}
-		data, err := json.Marshal(msg)
+		data, err := generationFunc()
 		if err != nil {
-			return nil, err
+			return err
 		}
-		return data, nil
-	case "command":
-		msg, err := generateCommandMessage(yggdrasil.MessageType(messageType), responseTo, version, []byte(content))
-		if err != nil {
-			return nil, err
-		}
-		data, err := json.Marshal(msg)
-		if err != nil {
-			return nil, err
-		}
-		return data, nil
-	default:
-		return nil, fmt.Errorf("unsupported message type: %v", messageType)
+		fmt.Println(data)
+		return nil
 	}
+
+	return cli.ShowAppHelp(c)
 }
